rfctr loadTextureFileWorker, use SharedArrayBuffer
