+++ src/utils/color-conversions/index.ts
@@ -6,3 +6,4 @@ export { default as rgb565ToRgba8888 } from './rgb565ToRgba8888';
 export { default as rgbaToArgb1555 } from './rgbaToArgb1555';
 export { default as rgbaToArgb4444 } from './rgbaToArgb4444';
 export { default as rgbaToRgb565 } from './rgbaToRgb565';
+export { default as rgba8888TargetOps } from './rgba8888TargetOps';
